+++ cmd/duel-masters/main.go
@@ -5,6 +5,7 @@ import (
 	"math/rand"
 	"os"
 	"strconv"
+	"strings"
 	"time"
 
 	"duel-masters/api"
@@ -24,8 +25,25 @@ func init() {
 		logrus.Warn("Failed to load .env file")
 	}
 
-	logrus.SetFormatter(&logrus.JSONFormatter{})
-	logrus.SetLevel(logrus.DebugLevel)
+	logrus.SetFormatter(&logrus.TextFormatter{})
+
+	loglevel := strings.ToLower(os.Getenv("log_level"))
+
+	switch loglevel {
+	case "debug":
+		logrus.SetLevel(logrus.DebugLevel)
+	case "info":
+		logrus.SetLevel(logrus.InfoLevel)
+	case "warn":
+		logrus.SetLevel(logrus.WarnLevel)
+	case "error":
+		logrus.SetLevel(logrus.ErrorLevel)
+	default:
+		loglevel = "info"
+		logrus.SetLevel(logrus.InfoLevel)
+	}
+
+	logrus.Infof("Using log level %s", loglevel)
 
 	rand.Seed(time.Now().UnixNano())
 }