+++ game/cards/dm01/ghost.go
@@ -49,7 +49,7 @@ func MaskedHorrorShadowOfScorn(c *match.Card) {
 					return
 				}
 
-				discardedCard, err := ctx.Match.Opponent(card.Player).MoveCard(hand[rand.Intn(len(hand))].ID, match.HAND, match.GRAVEYARD)
+				discardedCard, err := ctx.Match.Opponent(card.Player).MoveCard(hand[rand.Intn(len(hand))].ID, match.HAND, match.GRAVEYARD, card.ID)
 				if err == nil {
 					ctx.Match.Chat("Server", fmt.Sprintf("%s was discarded from %s's hand", discardedCard.Name, discardedCard.Player.Username()))
 				}