+++ game/cards/dm02/guardian.go
@@ -49,7 +49,7 @@ func PhalEegaDawnGuardian(c *match.Card) {
 			if x.ID == card.ID {
 				return
 			}
-			card.Player.MoveCard(x.ID, match.GRAVEYARD, match.HAND)
+			card.Player.MoveCard(x.ID, match.GRAVEYARD, match.HAND, card.ID)
 			ctx.Match.Chat("Server", fmt.Sprintf("%s was returned to %s's hand from their graveyard by Phal Eega, Dawn Guardian", x.Name, x.Player.Username()))
 		})