+++ game/cards/dm03/liquid_people.go
@@ -27,7 +27,7 @@ func AquaDeformer(c *match.Card) {
 				cards := match.Search(card.Player, ctx.Match, card.Player, match.MANAZONE, "Aqua Deformer: Select 2 cards from your manazone that will be sent to your hand", 2, 2, false)
 
 				for _, crd := range cards {
-					card.Player.MoveCard(crd.ID, match.MANAZONE, match.HAND)
+					card.Player.MoveCard(crd.ID, match.MANAZONE, match.HAND, card.ID)
 					ctx.Match.Chat("Server", fmt.Sprintf("%s was moved to %s's hand from their mana zone", crd.Name, ctx.Match.PlayerRef(card.Player).Socket.User.Username))
 				}
 
@@ -37,7 +37,7 @@ func AquaDeformer(c *match.Card) {
 				opponentCards := match.Search(ctx.Match.Opponent(card.Player), ctx.Match, ctx.Match.Opponent(card.Player), match.MANAZONE, "Aqua Deformer: Select 2 cards from your manazone that will be sent to your hand", 2, 2, false)
 
 				for _, crd := range opponentCards {
-					ctx.Match.Opponent(card.Player).MoveCard(crd.ID, match.MANAZONE, match.HAND)
+					ctx.Match.Opponent(card.Player).MoveCard(crd.ID, match.MANAZONE, match.HAND, card.ID)
 					ctx.Match.Chat("Server", fmt.Sprintf("%s was moved to %s's hand from their mana zone", crd.Name, ctx.Match.PlayerRef(ctx.Match.Opponent(card.Player)).Socket.User.Username))
 				}